backend/requirements.txt
Flask
flask-cors
Werkzeug
pytesseract
Pillow
concurrent.futures
requests
libreoffice
imagemagick
